Refactor player input into action facade
