+++ src/arena/views/debug.rs
@@ -7,7 +7,7 @@ use specs::prelude::*;
 use super::{screen_rect_for_map_grid, View};
 use crate::after_image::{FontColor, FontSize, RenderCanvas, TextRenderer};
 use crate::arena::components::*;
-use crate::arena::read_state;
+use crate::arena::battle_actions;
 use crate::atlas::BoxResult;
 use crate::clash::{MapComponent, Point};
 
@@ -26,7 +26,7 @@ impl<'a> DebugView<'a> {
 
 impl<'a> View for DebugView<'a> {
     fn render(&self, ecs: &World, canvas: &mut RenderCanvas, _frame: u64) -> BoxResult<()> {
-        if let BattleSceneState::Debug(kind) = read_state(&ecs) {
+        if let BattleSceneState::Debug(kind) = battle_actions::read_state(&ecs) {
             let state = format!("Debug: {}", kind.to_string());
             self.text
                 .render_text(&state, self.position.x, self.position.y, canvas, FontSize::Small, FontColor::Red)?;
